refactor: modal logic
